move solana validator files to ./blockchain
