+++ blockchain/solana/keypair/id3.json
@@ -1,0 +1,1 @@
+[111,197,242,14,118,44,66,203,60,194,40,110,92,232,238,66,8,252,188,88,11,83,103,105,48,172,125,30,146,126,150,226,209,157,201,11,117,149,72,146,24,180,136,1,49,113,203,216,165,60,57,180,135,207,103,204,255,56,145,144,182,123,224,245]
