+++ lib/sol/SolRpc.js
@@ -82,6 +82,10 @@ class SolRPC {
       address = new Web3.PublicKey(address);
       const block = await this.getTip();
 
+      if (!(fromAccountKeypair instanceof Web3.Keypair)) {
+        throw new Error('Invalid Solana Keypair object');
+      }
+
       // versioned tx
       if (txType == 0) {
         const instructions = [
